1,修改cqssc 的qszx 玩法 2,修改开奖功能 3,修改cqssc 获取开奖号码

+++ src/main/java/com/jll/common/utils/Utils.java
@@ -45,7 +45,7 @@ public class Utils {
 		Random random = new Random();		
 		
 		ret.append(dateFormat.format(nowTime));
-		ret.append(numFormat.format(seq));
+		ret.append(numFormat.format(seq.intValue()));
 		for(int i = 0; i < 4; i++) {
 			int currIndex = random.nextInt(10);
 			ret.append(Integer.toString(currIndex));
